backend/requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
bitarray==3.0.0
certifi==2024.12.14
charset-normalizer==3.4.1
ckzg==2.0.1
cytoolz==1.0.1
Django==5.1.4
django-cors-headers==4.6.0
django-cron==0.6.0
django-ratelimit==4.1.0
django-redis==5.4.0
django-secure==1.0.2
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
eth-account==0.13.4
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.6.1
eth-rlp==2.1.0
eth-typing==5.1.0
eth-utils==5.1.0
eth_abi==5.2.0
hexbytes==1.3.0
idna==3.10
itsdangerous==2.2.0
parsimonious==0.10.0
pycryptodome==3.21.0
pydantic==2.10.5
pydantic_core==2.27.2
PyJWT==2.10.1
redis==5.2.1
regex==2024.11.6
requests==2.32.3
rlp==4.0.1
sqlparse==0.5.3
toolz==1.0.0
typing_extensions==4.12.2
urllib3==2.3.0
